Lambda implementation
This is a go implementation for the kittenbot lambda. We bundle the
binary into a Docker image which will be our custom lambda runtime.

The actual implementation is pretty simple:
- Do some setup that might error. We do this early to avoid generating
  an image if something goes wrong here.
- Call out to dezgo to generate an image.
- Copy the generated image to our S3 site bucket.
- Generate a new index.html pointing to our new image plus embedding its
  metadata.
- Upload the new index.html to our S3 site bucket.
- Copy the new index.html to the date stamped html.
- Create a CloudFront invalidation for index.html so the new version
  will be served immediately.

Configuration is done via environment variables:
- PROMPT: Passed on to the dezgo as is.
- MODEL: A valid AI model in dezgo.
- BUCKET: S3 site bucket.
- DISTRIBUTION: CloudFront distribution.
- DEZGO_KEY: For local dev. The actual dezgo API key.
- DEZGO_KEY_PARAM: Path of SSM parameter containing the dezgo API key.

index.html contains meta tags for the prompt, model, and generated image
seed. These tags are also included as custom tags on the S3 objects;
those custom tags get passed on by CloudFront as headers.