Generate html on the fly
This is a big change. Instead of generating static html alongside the
image we point Cloudfront at a S3 Object Lambda Access Point for html.
If a page isn't in the Cloudfront cache it will try to pull it from the
lambda access point which will invoke the html lambda to generate the
page content.

This does mean we'll possibly be generating a page multiple times since
a page will eventually be evicted from the cache. However, this allows
us to update the html template and older pages will automatically get
the changes. :)